docs/easysite/bot/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.28.0
apscheduler>=3.10.4
